backend/requirements.txt
# Flask and Web Framework
Flask==3.0.0
Flask-CORS==4.0.0
Flask-PyMongo==2.3.0
python-dotenv==1.0.0

# Database
pymongo==4.6.1

# LLM and AI
langchain>=0.1.0
langchain-community>=0.0.20
langchain-core>=0.1.0
langchain-openai>=0.1.0
langsmith>=0.0.77
openai>=1.12.0
chromadb==0.4.22
sentence-transformers==2.3.1
huggingface-hub>=0.20.0

# Code Analysis
gitpython==3.1.40
pygments==2.17.2
tree-sitter==0.20.4

# GitHub/GitLab Integration
PyGithub==2.1.1
python-gitlab==4.4.0
requests==2.31.0

# Security and Authentication
PyJWT==2.8.0
cryptography==41.0.7

# Utilities
pydantic==2.5.3
pyyaml==6.0.1
click==8.1.7

# Testing
pytest==7.4.3
pytest-flask==1.3.0
pytest-cov==4.1.0
pytest-mock==3.12.0

# Code Quality
black==23.12.1
flake8==7.0.0
pylint==3.0.3

# Monitoring and Logging
python-json-logger==2.0.7

# Task Queue (optional for async processing)
celery==5.3.4
redis==5.0.1
